Issues 356 - Refactoring FileUtils
